distributed functions from ctf_utils to core
