+++ src/core/ctf.h
@@ -1,3 +1,5 @@
+class Image;
+
 class CTF {
 
   private:
@@ -168,4 +170,9 @@ class CTF {
     float ReturnPhaseAberrationMaximum( );
     float ReturnPhaseAberrationMaximumGivenDefocus(float defocus);
     float ReturnAzimuthToUseFor1DPlots( );
+    void  ComputeImagesWithNumberOfExtremaAndCTFValues(Image* number_of_extrema, Image* ctf_values);
 };
+
+void Renormalize1DSpectrumForFRC(int number_of_bins, double average[], double fit[], float number_of_extrema_profile[]);
+void ComputeFRCBetween1DSpectrumAndFit(int number_of_bins, double average[], double fit[], float number_of_extrema_profile[], double frc[], double frc_sigma[], int first_fit_bin);
+int  ReturnSpectrumBinNumber(int number_of_bins, float number_of_extrema_profile[], Image* number_of_extrema, long address, Image* ctf_values, float ctf_values_profile[]);
