--- src/programs/ctffind/ctffind_utils.cpp
@@ -1,639 +1,0 @@
-#include "../../core/core_headers.h"
-#include "./ctffind.h"
-
-/*
- * Stretch or shrink the powerspectrum to maintain the same bix size, but represent a new physical pixel size (target_pixel_size_after_resampling).
- * Returns the actual new pixel size, which might be sligthly different.
- */
-float DilatePowerspectrumToNewPixelSize(bool resample_if_pixel_too_small, float pixel_size_of_input_image, float target_pixel_size_after_resampling,
-                                        int box_size, Image* current_power_spectrum, Image* resampled_power_spectrum, bool do_resampling, float stretch_factor) {
-    int   temporary_box_size;
-    int   stretched_dimension;
-    float pixel_size_for_fitting;
-    bool  resampling_is_necessary;
-
-    Image temp_image;
-
-    // Resample the amplitude spectrum
-    if ( resample_if_pixel_too_small && pixel_size_of_input_image < target_pixel_size_after_resampling ) {
-        // The input pixel was too small, so let's resample the amplitude spectrum into a large temporary box, before clipping the center out for fitting
-        temporary_box_size = round(float(box_size) / pixel_size_of_input_image * target_pixel_size_after_resampling);
-        if ( IsOdd(temporary_box_size) )
-            temporary_box_size++;
-        resampling_is_necessary = current_power_spectrum->logical_x_dimension != box_size || current_power_spectrum->logical_y_dimension != box_size;
-        if ( do_resampling ) {
-            if ( resampling_is_necessary || stretch_factor != 1.0f ) {
-                stretched_dimension = myroundint(temporary_box_size * stretch_factor);
-                if ( IsOdd(stretched_dimension) )
-                    stretched_dimension++;
-                if ( fabsf(stretched_dimension - temporary_box_size * stretch_factor) > fabsf(stretched_dimension - 2 - temporary_box_size * stretch_factor) )
-                    stretched_dimension -= 2;
-
-                current_power_spectrum->ForwardFFT(false);
-                resampled_power_spectrum->Allocate(stretched_dimension, stretched_dimension, 1, false);
-                current_power_spectrum->ClipInto(resampled_power_spectrum);
-                resampled_power_spectrum->BackwardFFT( );
-                temp_image.Allocate(box_size, box_size, 1, true);
-                temp_image.SetToConstant(0.0); // To avoid valgrind uninitialised errors, but maybe this is a waste?
-                resampled_power_spectrum->ClipInto(&temp_image);
-                resampled_power_spectrum->Consume(&temp_image);
-            }
-            else {
-                resampled_power_spectrum->CopyFrom(current_power_spectrum);
-            }
-        }
-        pixel_size_for_fitting = pixel_size_of_input_image * float(temporary_box_size) / float(box_size);
-    }
-    else {
-        // The regular way (the input pixel size was large enough)
-        resampling_is_necessary = current_power_spectrum->logical_x_dimension != box_size || current_power_spectrum->logical_y_dimension != box_size;
-        if ( do_resampling ) {
-            if ( resampling_is_necessary || stretch_factor != 1.0f ) {
-                stretched_dimension = myroundint(box_size * stretch_factor);
-                if ( IsOdd(stretched_dimension) )
-                    stretched_dimension++;
-                if ( fabsf(stretched_dimension - box_size * stretch_factor) > fabsf(stretched_dimension - 2 - box_size * stretch_factor) )
-                    stretched_dimension -= 2;
-
-                current_power_spectrum->ForwardFFT(false);
-                resampled_power_spectrum->Allocate(stretched_dimension, stretched_dimension, 1, false);
-                current_power_spectrum->ClipInto(resampled_power_spectrum);
-                resampled_power_spectrum->BackwardFFT( );
-                temp_image.Allocate(box_size, box_size, 1, true);
-                temp_image.SetToConstant(0.0); // To avoid valgrind uninitialised errors, but maybe this is a waste?
-                resampled_power_spectrum->ClipInto(&temp_image);
-                resampled_power_spectrum->Consume(&temp_image);
-            }
-            else {
-                resampled_power_spectrum->CopyFrom(current_power_spectrum);
-            }
-        }
-        pixel_size_for_fitting = pixel_size_of_input_image;
-    }
-
-    return pixel_size_for_fitting;
-}
-
-/*
- * Go from an experimental radial average with decaying Thon rings to a function between 0.0 and 1.0 for every oscillation.
- * This is done by treating each interval between a zero and an extremum of the CTF separately, and for each of them,
- * sorting and ranking the values in the radial average.
- * Each value is then replaced by its rank, modified to make it looks like a |CTF| signal.
- * This makes sense as a preparation for evaluating the quality of fit of a CTF when we want to ignore the amplitude of the Thon
- * rings and just focus on whether the fit agrees in terms of the positions of the zeros and extrema.
- * Without this, a very good fit doesn't always have a great FRC for regions where the experimental radial average is decaying rapidly.
- */
-void Renormalize1DSpectrumForFRC(int number_of_bins, double average[], double fit[], float number_of_extrema_profile[]) {
-    int                 bin_counter;
-    int                 bin_of_previous_extremum;
-    int                 bin_of_current_extremum;
-    int                 i;
-    int                 bin_of_zero;
-    std::vector<float>  temp_vector;
-    std::vector<size_t> temp_ranks;
-    float               number_of_extrema_delta;
-    //
-    bin_of_previous_extremum = 0;
-    bin_of_current_extremum  = 0;
-    for ( bin_counter = 1; bin_counter < number_of_bins; bin_counter++ ) {
-        number_of_extrema_delta = number_of_extrema_profile[bin_counter] - number_of_extrema_profile[bin_counter - 1];
-        if ( number_of_extrema_delta >= 0.9 && number_of_extrema_delta <= 1.9 ) // if the CTF is oscillating too quickly, let's not do anything
-        {
-            // We just passed an extremum, at bin_counter-1
-            // (number_of_extrema_profile keeps track of the count of extrema before the spatial frequency corresponding to this bin)
-            bin_of_current_extremum = bin_counter - 1;
-            if ( bin_of_previous_extremum > 0 ) {
-                if ( (bin_of_current_extremum - bin_of_previous_extremum >= 4 && false) || (number_of_extrema_profile[bin_counter] < 7) ) {
-                    // Loop from the previous extremum to the one we just found
-                    // (there is a zero in between, let's find it)
-                    // TODO: redefine the zero as the lowest point between the two extrema?
-                    bin_of_zero = (bin_of_current_extremum - bin_of_previous_extremum) / 2 + bin_of_previous_extremum;
-                    for ( i = bin_of_previous_extremum; i < bin_of_current_extremum; i++ ) {
-                        if ( fit[i] < fit[i - 1] && fit[i] < fit[i + 1] )
-                            bin_of_zero = i;
-                    }
-                    //wxPrintf("bin zero = %i\n",bin_of_zero);
-
-                    // Now we can rank before the zero (the downslope)
-                    //wxPrintf("downslope (including zero)...\n");
-                    temp_vector.clear( );
-                    for ( i = bin_of_previous_extremum; i <= bin_of_zero; i++ ) {
-                        //wxPrintf("about to push back %f\n",float(average[i]));
-                        temp_vector.push_back(float(average[i]));
-                    }
-                    temp_ranks = rankSort(temp_vector);
-                    for ( i = bin_of_previous_extremum; i <= bin_of_zero; i++ ) {
-                        //wxPrintf("replaced %f",average[i]);
-                        average[i] = double(float(temp_ranks.at(i - bin_of_previous_extremum)) / float(temp_vector.size( ) - 1));
-                        average[i] = sin(average[i] * PI * 0.5);
-                        //wxPrintf(" with %f\n",average[i]);
-                    }
-
-                    // Now we can rank after the zero (upslope)
-                    //wxPrintf("upslope...\n");
-                    temp_vector.clear( );
-                    for ( i = bin_of_zero + 1; i < bin_of_current_extremum; i++ ) {
-                        //wxPrintf("about to push back %f\n",float(average[i]));
-                        temp_vector.push_back(float(average[i]));
-                    }
-                    temp_ranks = rankSort(temp_vector);
-                    for ( i = bin_of_zero + 1; i < bin_of_current_extremum; i++ ) {
-                        //wxPrintf("[rank]bin %i: replaced %f",i,average[i]);
-                        average[i] = double(float(temp_ranks.at(i - bin_of_zero - 1) + 1) / float(temp_vector.size( ) + 1));
-                        average[i] = sin(average[i] * PI * 0.5);
-                        //wxPrintf(" with %f\n",average[i]);
-                    }
-                    //MyDebugAssertTrue(abs(average[bin_of_zero]) < 0.01,"Zero bin (%i) isn't set to zero: %f\n", bin_of_zero, average[bin_of_zero]);
-                }
-                else {
-                    // A simpler way, without ranking, is just normalize
-                    // between 0.0 and 1.0 (this usually works quite well when Thon rings are on a flat background anyway)
-                    float min_value = 1.0;
-                    float max_value = 0.0;
-                    for ( i = bin_of_previous_extremum; i < bin_of_current_extremum; i++ ) {
-                        if ( average[i] > max_value )
-                            max_value = average[i];
-                        if ( average[i] < min_value )
-                            min_value = average[i];
-                    }
-                    for ( i = bin_of_previous_extremum; i < bin_of_current_extremum; i++ ) {
-                        //wxPrintf("bin %i: replaced %f",i,average[i]);
-                        average[i] -= min_value;
-                        if ( max_value - min_value > 0.0001 )
-                            average[i] /= (max_value - min_value);
-                        //wxPrintf(" with %f\n",average[i]);
-                    }
-                }
-            }
-            bin_of_previous_extremum = bin_of_current_extremum;
-        }
-        MyDebugAssertFalse(std::isnan(average[bin_counter]), "Average is NaN for bin %i\n", bin_counter);
-    }
-}
-
-//
-void ComputeFRCBetween1DSpectrumAndFit(int number_of_bins, double average[], double fit[], float number_of_extrema_profile[], double frc[], double frc_sigma[], int first_fit_bin) {
-
-    MyDebugAssertTrue(first_fit_bin >= 0, "Bad first fit bin on entry: %i", first_fit_bin);
-
-    int    bin_counter;
-    int    half_window_width[number_of_bins];
-    int    bin_of_previous_extremum;
-    int    i;
-    int    first_bin, last_bin;
-    double spectrum_mean, fit_mean;
-    double spectrum_sigma, fit_sigma;
-    double cross_product;
-    float  number_of_bins_in_window;
-
-    const int minimum_window_half_width = number_of_bins / 40;
-
-    // First, work out the size of the window over which we'll compute the FRC value
-    bin_of_previous_extremum = 0;
-    for ( bin_counter = 1; bin_counter < number_of_bins; bin_counter++ ) {
-        if ( number_of_extrema_profile[bin_counter] != number_of_extrema_profile[bin_counter - 1] ) {
-            for ( i = bin_of_previous_extremum; i < bin_counter; i++ ) {
-                half_window_width[i] = std::max(minimum_window_half_width, int((1.0 + 0.1 * float(number_of_extrema_profile[bin_counter])) * float(bin_counter - bin_of_previous_extremum + 1)));
-                half_window_width[i] = std::min(half_window_width[i], number_of_bins / 2 - 1);
-                MyDebugAssertTrue(half_window_width[i] < number_of_bins / 2, "Bad half window width: %i. Number of bins: %i\n", half_window_width[i], number_of_bins);
-            }
-            bin_of_previous_extremum = bin_counter;
-        }
-    }
-    half_window_width[0] = half_window_width[1];
-    for ( bin_counter = bin_of_previous_extremum; bin_counter < number_of_bins; bin_counter++ ) {
-        half_window_width[bin_counter] = half_window_width[bin_of_previous_extremum - 1];
-    }
-
-    // Now compute the FRC for each bin
-    for ( bin_counter = 0; bin_counter < number_of_bins; bin_counter++ ) {
-        if ( bin_counter < first_fit_bin ) {
-            frc[bin_counter]       = 1.0;
-            frc_sigma[bin_counter] = 0.0;
-        }
-        else {
-            spectrum_mean  = 0.0;
-            fit_mean       = 0.0;
-            spectrum_sigma = 0.0;
-            fit_sigma      = 0.0;
-            cross_product  = 0.0;
-            // Work out the boundaries
-            first_bin = bin_counter - half_window_width[bin_counter];
-            last_bin  = bin_counter + half_window_width[bin_counter];
-            if ( first_bin < first_fit_bin ) {
-                first_bin = first_fit_bin;
-                last_bin  = first_bin + 2 * half_window_width[bin_counter] + 1;
-            }
-            if ( last_bin >= number_of_bins ) {
-                last_bin  = number_of_bins - 1;
-                first_bin = last_bin - 2 * half_window_width[bin_counter] - 1;
-            }
-            MyDebugAssertTrue(first_bin >= 0 && first_bin < number_of_bins, "Bad first_bin: %i", first_bin);
-            MyDebugAssertTrue(last_bin >= 0 && last_bin < number_of_bins, "Bad last_bin: %i", last_bin);
-            // First pass
-            for ( i = first_bin; i <= last_bin; i++ ) {
-                spectrum_mean += average[i];
-                fit_mean += fit[i];
-            }
-            number_of_bins_in_window = float(2 * half_window_width[bin_counter] + 1);
-            //wxPrintf("bin %03i, number of extrema: %f, number of bins in window: %f , spectrum_sum = %f\n", bin_counter, number_of_extrema_profile[bin_counter], number_of_bins_in_window,spectrum_mean);
-            spectrum_mean /= number_of_bins_in_window;
-            fit_mean /= number_of_bins_in_window;
-            // Second pass
-            for ( i = first_bin; i <= last_bin; i++ ) {
-                cross_product += (average[i] - spectrum_mean) * (fit[i] - fit_mean);
-                spectrum_sigma += pow(average[i] - spectrum_mean, 2);
-                fit_sigma += pow(fit[i] - fit_mean, 2);
-            }
-            MyDebugAssertTrue(spectrum_sigma > 0.0 && spectrum_sigma < 10000.0, "Bad spectrum_sigma: %f\n", spectrum_sigma);
-            MyDebugAssertTrue(fit_sigma > 0.0 && fit_sigma < 10000.0, "Bad fit sigma: %f\n", fit_sigma);
-            if ( spectrum_sigma > 0.0 && fit_sigma > 0.0 ) {
-                frc[bin_counter] = cross_product / (sqrtf(spectrum_sigma / number_of_bins_in_window) * sqrtf(fit_sigma / number_of_bins_in_window)) / number_of_bins_in_window;
-            }
-            else {
-                frc[bin_counter] = 0.0;
-            }
-            frc_sigma[bin_counter] = 2.0 / sqrtf(number_of_bins_in_window);
-        }
-        //wxPrintf("First fit bin: %i\n", first_fit_bin);
-        MyDebugAssertTrue(frc[bin_counter] > -1.01 && frc[bin_counter] < 1.01, "Bad FRC value: %f", frc[bin_counter]);
-    }
-}
-
-// Rescale the spectrum and its 1D rotational avereage so that the peaks and troughs are at 0.0 and 1.0. The location of peaks and troughs are worked out
-// by parsing the suppilied 1D average_fit array
-void RescaleSpectrumAndRotationalAverage(Image* spectrum, Image* number_of_extrema, Image* ctf_values, int number_of_bins, double spatial_frequency[], double average[], double average_fit[], float number_of_extrema_profile[], float ctf_values_profile[], int last_bin_without_aliasing, int last_bin_with_good_fit) {
-    MyDebugAssertTrue(spectrum->is_in_memory, "Spectrum memory not allocated");
-    MyDebugAssertTrue(number_of_bins > 1, "Bad number of bins: %i\n", number_of_bins);
-
-    //
-    const bool spectrum_is_blank               = spectrum->IsConstant( );
-    const int  rescale_based_on_maximum_number = 2; // This peak will be used as a renormalization.
-    const int  sg_width                        = 7;
-    const int  sg_order                        = 2;
-    const bool rescale_peaks                   = false; // if this is false, only the background will be subtracted, the Thon rings "heights" will be unaffected
-    float      background[number_of_bins];
-    float      peak[number_of_bins];
-    int        bin_counter;
-    bool       at_a_maximum, at_a_minimum, maximum_at_previous_bin, minimum_at_previous_bin;
-    int        location_of_previous_maximum, location_of_previous_minimum;
-    int        current_maximum_number = 0;
-    int        normalisation_bin_number;
-    int        i;
-    int        j;
-    bool       actually_do_rescaling;
-    int        chosen_bin;
-    long       address;
-    int        last_bin_to_rescale;
-    float      min_scale_factor;
-    float      scale_factor;
-    float      rescale_peaks_to;
-
-    Curve* minima_curve = new Curve;
-    Curve* maxima_curve = new Curve;
-
-    // Initialise arrays and variables
-    for ( bin_counter = 0; bin_counter < number_of_bins; bin_counter++ ) {
-        background[bin_counter] = 0.0;
-        peak[bin_counter]       = 0.0;
-    }
-    location_of_previous_maximum = 0;
-    location_of_previous_minimum = 0;
-    current_maximum_number       = 0;
-    at_a_maximum                 = false;
-    at_a_minimum                 = true; // Note, this may not be true if we have the perfect phase plate
-
-    //
-    if ( ! spectrum_is_blank ) {
-        for ( bin_counter = 1; bin_counter < number_of_bins - 1; bin_counter++ ) {
-            // Remember where we were before - minimum, maximum or neither
-            maximum_at_previous_bin = at_a_maximum;
-            minimum_at_previous_bin = at_a_minimum;
-            // Are we at a CTF min or max?
-            at_a_minimum = (average_fit[bin_counter] <= average_fit[bin_counter - 1]) && (average_fit[bin_counter] <= average_fit[bin_counter + 1]);
-            at_a_maximum = (average_fit[bin_counter] >= average_fit[bin_counter - 1]) && (average_fit[bin_counter] >= average_fit[bin_counter + 1]);
-            // It could be that the CTF is constant in this region, in which case we stay at a minimum if we were there
-            if ( at_a_maximum && at_a_minimum ) {
-                at_a_minimum = minimum_at_previous_bin;
-                at_a_maximum = maximum_at_previous_bin;
-            }
-            // Fill in values for the background or peak by linear interpolation
-            if ( at_a_minimum ) {
-                for ( i = location_of_previous_minimum + 1; i <= bin_counter; i++ ) {
-                    // Linear interpolation of average values at the peaks and troughs of the CTF
-                    background[i] = average[location_of_previous_minimum] * float(bin_counter - i) / float(bin_counter - location_of_previous_minimum) + average[bin_counter] * float(i - location_of_previous_minimum) / float(bin_counter - location_of_previous_minimum);
-                }
-                location_of_previous_minimum = bin_counter;
-                minima_curve->AddPoint(spatial_frequency[bin_counter], average[bin_counter]);
-            }
-            if ( at_a_maximum ) {
-                if ( (! maximum_at_previous_bin) && (average_fit[bin_counter] > 0.7) )
-                    current_maximum_number = current_maximum_number + 1;
-                for ( i = location_of_previous_maximum + 1; i <= bin_counter; i++ ) {
-                    // Linear interpolation of average values at the peaks and troughs of the CTF
-                    peak[i] = average[location_of_previous_maximum] * float(bin_counter - i) / float(bin_counter - location_of_previous_maximum) + average[bin_counter] * float(i - location_of_previous_maximum) / float(bin_counter - location_of_previous_maximum);
-                    //
-                    if ( current_maximum_number == rescale_based_on_maximum_number )
-                        normalisation_bin_number = bin_counter;
-                }
-                location_of_previous_maximum = bin_counter;
-                maxima_curve->AddPoint(spatial_frequency[bin_counter], average[bin_counter]);
-            }
-            if ( at_a_maximum && at_a_minimum ) {
-                MyPrintfRed("Rescale spectrum: Error. At a minimum and a maximum simultaneously.");
-                //TODO: return false instead
-                DEBUG_ABORT;
-            }
-        }
-
-        // Fit the minima and maximum curves using Savitzky-Golay smoothing
-        if ( maxima_curve->number_of_points > sg_width )
-            maxima_curve->FitSavitzkyGolayToData(sg_width, sg_order);
-        if ( minima_curve->number_of_points > sg_width )
-            minima_curve->FitSavitzkyGolayToData(sg_width, sg_order);
-
-        // Replace the background and peak envelopes with the smooth min/max curves
-        for ( bin_counter = 0; bin_counter < number_of_bins; bin_counter++ ) {
-            if ( minima_curve->number_of_points > sg_width )
-                background[bin_counter] = minima_curve->ReturnSavitzkyGolayInterpolationFromX(spatial_frequency[bin_counter]);
-            if ( maxima_curve->number_of_points > sg_width )
-                peak[bin_counter] = maxima_curve->ReturnSavitzkyGolayInterpolationFromX(spatial_frequency[bin_counter]);
-        }
-
-        // Now that we have worked out a background and a peak envelope, let's do the actual rescaling
-        actually_do_rescaling = (peak[normalisation_bin_number] - background[normalisation_bin_number]) > 0.0;
-        if ( last_bin_without_aliasing != 0 ) {
-            last_bin_to_rescale = std::min(last_bin_with_good_fit, last_bin_without_aliasing);
-        }
-        else {
-            last_bin_to_rescale = last_bin_with_good_fit;
-        }
-        if ( actually_do_rescaling ) {
-            min_scale_factor = 0.2;
-            rescale_peaks_to = 0.75;
-            address          = 0;
-            for ( j = 0; j < spectrum->logical_y_dimension; j++ ) {
-                for ( i = 0; i < spectrum->logical_x_dimension; i++ ) {
-                    chosen_bin = ReturnSpectrumBinNumber(number_of_bins, number_of_extrema_profile, number_of_extrema, address, ctf_values, ctf_values_profile);
-                    if ( chosen_bin >= 0 ) {
-                        if ( chosen_bin <= last_bin_to_rescale ) {
-                            spectrum->real_values[address] -= background[chosen_bin]; // This alone makes the spectrum look very nice already
-                            if ( rescale_peaks )
-                                spectrum->real_values[address] /= std::min(1.0f, std::max(min_scale_factor, peak[chosen_bin] - background[chosen_bin])) / rescale_peaks_to; // This is supposed to help "boost" weak Thon rings
-                        }
-                        else {
-                            spectrum->real_values[address] -= background[last_bin_to_rescale];
-                            if ( rescale_peaks )
-                                spectrum->real_values[address] /= std::min(1.0f, std::max(min_scale_factor, peak[last_bin_to_rescale] - background[last_bin_to_rescale])) / rescale_peaks_to;
-                        }
-                    }
-                    else {
-                        //TODO: return false
-                    }
-                    //
-                    address++;
-                }
-                address += spectrum->padding_jump_value;
-            }
-        }
-        else {
-            MyDebugPrint("(RescaleSpectrumAndRotationalAverage) Warning: bad peak/background detection");
-        }
-
-        // Rescale the 1D average
-        if ( peak[normalisation_bin_number] > background[normalisation_bin_number] ) {
-            for ( bin_counter = 0; bin_counter < number_of_bins; bin_counter++ ) {
-
-                average[bin_counter] = (average[bin_counter] - background[bin_counter]) / (peak[normalisation_bin_number] - background[normalisation_bin_number]) * 0.95;
-                // We want peaks to reach at least 0.1
-                if ( ((peak[bin_counter] - background[bin_counter]) < 0.1) && (fabs(peak[bin_counter] - background[bin_counter]) > 0.000001) && bin_counter <= last_bin_without_aliasing ) {
-                    average[bin_counter] = average[bin_counter] / (peak[bin_counter] - background[bin_counter]) * (peak[normalisation_bin_number] - background[normalisation_bin_number]) * 0.1;
-                }
-            }
-        }
-        else {
-            MyDebugPrint("(RescaleSpectrumAndRotationalAverage): unable to rescale 1D average experimental spectrum\n");
-        }
-
-    } // end of test of spectrum_is_blank
-
-    // Cleanup
-    delete minima_curve;
-    delete maxima_curve;
-}
-
-/*
- * Compute average value in power spectrum as a function of wave function aberration. This allows for averaging even when
- * there is significant astigmatism.
- * This should be nicer than counting zeros and looking for nearest CTF value as described in the original ctffind4 manuscript.
- * Inspired by gctf and others, but I think more robust because it takes into account that the aberration decreases again at
- * very high spatial frequencies, when Cs takes over from defocus.
- */
-void ComputeEquiPhaseAverageOfPowerSpectrum(Image* spectrum, CTF* ctf, Curve* epa_pre_max, Curve* epa_post_max) {
-    MyDebugAssertTrue(spectrum->is_in_memory, "Spectrum memory not allocated");
-
-    const bool spectrum_is_blank = spectrum->IsConstant( );
-
-    const int  curve_oversampling_factor = 3;
-    const bool curve_x_is_linear         = true;
-
-    /*
-	 * Initialize the curve objects. One keeps track of EPA pre phase aberration maximum (before Cs term takes over), the other post.
-	 * In the case where we are overfocus (negative defocus value), the phase aberration starts at 0.0 at the origin
-	 * and just gets more and more negative
-	 *
-	 * This is one of the messiest parts of the code. I really need to come up with a cleaner way to decide how many points
-	 * to give each curve. This is a goldilocks problem: too few or too many both give worse curves and FRCs.
-	 */
-    if ( curve_x_is_linear ) {
-        float maximum_aberration_in_ctf            = ctf->ReturnPhaseAberrationMaximum( );
-        float maximum_sq_freq_in_spectrum          = powf(spectrum->fourier_voxel_size_x * spectrum->logical_lower_bound_complex_x, 2) + powf(spectrum->fourier_voxel_size_y * spectrum->logical_lower_bound_complex_y, 2);
-        float lowest_sq_freq_of_ctf_aberration_max = std::min(fabs(ctf->ReturnSquaredSpatialFrequencyOfPhaseShiftExtremumGivenDefocus(ctf->GetDefocus1( ))),
-                                                              fabs(ctf->ReturnSquaredSpatialFrequencyOfPhaseShiftExtremumGivenDefocus(ctf->GetDefocus2( ))));
-
-        float maximum_abs_aberration_in_spectrum = std::max(fabs(ctf->PhaseShiftGivenSquaredSpatialFrequencyAndDefocus(maximum_sq_freq_in_spectrum, ctf->GetDefocus1( ))),
-                                                            fabs(ctf->PhaseShiftGivenSquaredSpatialFrequencyAndDefocus(maximum_sq_freq_in_spectrum, ctf->GetDefocus2( ))));
-
-        /*
-		 * Minimum phase aberration might be 0.0 + additional_phase_shift (at the origin), or if the phase aberration function
-		 * peaks before Nyquist, it might be at the edge of the spectrum
-		 */
-        float minimum_aberration_in_ctf_at_edges = std::min(ctf->PhaseShiftGivenSquaredSpatialFrequencyAndDefocus(maximum_sq_freq_in_spectrum, ctf->GetDefocus1( )),
-                                                            ctf->PhaseShiftGivenSquaredSpatialFrequencyAndDefocus(maximum_sq_freq_in_spectrum, ctf->GetDefocus2( )));
-
-        // Watch out: messy heuristics
-        int number_of_points_pre_max  = std::max(2, myroundint(spectrum->ReturnMaximumDiagonalRadius( ) * curve_oversampling_factor * maximum_aberration_in_ctf / maximum_abs_aberration_in_spectrum));
-        int number_of_points_post_max = std::max(2, myroundint(spectrum->ReturnMaximumDiagonalRadius( ) * curve_oversampling_factor));
-
-        epa_pre_max->SetupXAxis(ctf->GetAdditionalPhaseShift( ), maximum_aberration_in_ctf, number_of_points_pre_max);
-        epa_post_max->SetupXAxis(std::min(maximum_aberration_in_ctf, minimum_aberration_in_ctf_at_edges - 0.5f * fabsf(minimum_aberration_in_ctf_at_edges)), maximum_aberration_in_ctf, number_of_points_post_max);
-    }
-    else {
-        MyDebugAssertTrue(false, "Not implemented");
-    }
-    epa_pre_max->SetYToConstant(0.0);
-    epa_post_max->SetYToConstant(0.0);
-
-    /*
-	 * We'll also need to keep track of the number of values
-	 */
-    Curve* count_pre_max  = new Curve;
-    Curve* count_post_max = new Curve;
-    count_pre_max->CopyFrom(epa_pre_max);
-    count_post_max->CopyFrom(epa_post_max);
-
-    if ( ! spectrum_is_blank ) {
-        long  address = 0;
-        int   i, j;
-        float i_logi, j_logi;
-        float i_logi_sq, j_logi_sq;
-        float current_spatial_frequency_squared;
-        float current_azimuth;
-        float current_phase_aberration;
-        float sq_sf_of_phase_aberration_maximum;
-        float current_defocus;
-        for ( j = 0; j < spectrum->logical_y_dimension; j++ ) {
-            j_logi    = float(j - spectrum->physical_address_of_box_center_y) * spectrum->fourier_voxel_size_y;
-            j_logi_sq = powf(j_logi, 2);
-            for ( i = 0; i < spectrum->logical_x_dimension; i++ ) {
-                i_logi    = float(i - spectrum->physical_address_of_box_center_x) * spectrum->fourier_voxel_size_x;
-                i_logi_sq = powf(i_logi, 2);
-                //
-                current_spatial_frequency_squared = j_logi_sq + i_logi_sq;
-                current_azimuth                   = atan2(j_logi, i_logi);
-                current_defocus                   = ctf->DefocusGivenAzimuth(current_azimuth);
-                current_phase_aberration          = ctf->PhaseShiftGivenSquaredSpatialFrequencyAndDefocus(current_spatial_frequency_squared, current_defocus);
-                //
-                sq_sf_of_phase_aberration_maximum = ctf->ReturnSquaredSpatialFrequencyOfPhaseShiftExtremumGivenDefocus(current_defocus);
-                //
-                if ( current_spatial_frequency_squared <= sq_sf_of_phase_aberration_maximum ) {
-                    // Add to pre-max
-                    epa_pre_max->AddValueAtXUsingLinearInterpolation(current_phase_aberration, spectrum->real_values[address], curve_x_is_linear);
-                    count_pre_max->AddValueAtXUsingLinearInterpolation(current_phase_aberration, 1.0, curve_x_is_linear);
-                }
-                else {
-                    /*
-					 * We are after the maximum phase aberration (i.e. the Cs term has taken over, phase aberration is decreasing as a function of sf)
-					 */
-                    // Add to post-max
-                    epa_post_max->AddValueAtXUsingLinearInterpolation(current_phase_aberration, spectrum->real_values[address], curve_x_is_linear);
-                    count_post_max->AddValueAtXUsingLinearInterpolation(current_phase_aberration, 1.0, curve_x_is_linear);
-                }
-                //
-                address++;
-            }
-            address += spectrum->padding_jump_value;
-        }
-
-        /*
-		 * Do the averaging
-		 */
-        epa_pre_max->DivideBy(count_pre_max);
-        epa_post_max->DivideBy(count_post_max);
-    }
-
-    delete count_pre_max;
-    delete count_post_max;
-}
-
-//
-
-/*
-integer function ComputePowerSpectrumBinNumber(number_of_bins,number_of_extrema_profile,number_of_extrema, &
-                                                i,j,ctf_values,ctf_values_profile) result(chosen_bin)
-    integer,        intent(in)  ::  number_of_bins
-    real,           intent(in)  ::  number_of_extrema_profile(:)
-    type(Image),    intent(in)  ::  number_of_extrema
-    integer,        intent(in)  ::  i,j                         !<  Physical memory address
-    type(Image),    intent(in)  ::  ctf_values
-    real,           intent(in)  ::  ctf_values_profile(:)
-    ! private variables
-    integer     ::  current_bin
-    real        ::  diff_number_of_extrema, diff_number_of_extrema_previous, diff_number_of_extrema_next
-    real        ::  ctf_diff_from_current_bin
-    real        ::  ctf_diff_from_current_bin_old
-    ! Let's find the bin which has the same number of preceding extrema and the most similar ctf value
-    ctf_diff_from_current_bin = huge(1.0e0)
-    chosen_bin = 0
-    do current_bin=1,number_of_bins
-        diff_number_of_extrema  = abs(number_of_extrema%real_values(i,j,1) - number_of_extrema_profile(current_bin))
-        if (current_bin .gt. 1) then
-            diff_number_of_extrema_previous = abs(number_of_extrema%real_values(i,j,1) &
-                                                - number_of_extrema_profile(current_bin-1))
-        else
-            diff_number_of_extrema_previous = huge(1.0e0)
-        endif
-        if (current_bin .lt. number_of_bins) then
-            diff_number_of_extrema_next     = abs(number_of_extrema%real_values(i,j,1) &
-                                                - number_of_extrema_profile(current_bin+1))
-        else
-            diff_number_of_extrema_next = huge(1.0e0)
-        endif
-        if (number_of_extrema%real_values(i,j,1) .gt. number_of_extrema_profile(number_of_bins)) then
-            chosen_bin = number_of_bins
-        else
-            if (        diff_number_of_extrema .le. 0.01 &
-                .or.    (     diff_number_of_extrema .lt. diff_number_of_extrema_previous &
-                        .and. diff_number_of_extrema .le. diff_number_of_extrema_next &
-                        .and. number_of_extrema_profile(max(current_bin-1,1)) &
-                            .ne. number_of_extrema_profile(min(current_bin+1,number_of_bins))) &
-                ) then
-                ! We're nearly there
-                ! Let's look for the position of the nearest CTF value
-                ctf_diff_from_current_bin_old = ctf_diff_from_current_bin
-                ctf_diff_from_current_bin = abs(ctf_values%real_values(i,j,1) - ctf_values_profile(current_bin))
-                if (ctf_diff_from_current_bin .lt. ctf_diff_from_current_bin_old) then
-                    chosen_bin = current_bin
-                endif
-            endif
-        endif
-    enddo
-    if (chosen_bin .eq. 0) then
-        print *, number_of_extrema_profile
-        print *, i, j, number_of_extrema%real_values(i,j,1), ctf_values%real_values(i,j,1)
-        print *, diff_number_of_extrema, diff_number_of_extrema_previous, diff_number_of_extrema_next
-        call this_program%TerminateWithFatalError('ComputeRotationalAverageOfPowerSpectrum','Could not find bin')
-    endif
-end function ComputePowerSpectrumBinNumber
-*/
-
-// Compute an image where each pixel stores the number of preceding CTF extrema. This is described as image "E" in Rohou & Grigorieff 2015 (see Fig 3)
-void ComputeImagesWithNumberOfExtremaAndCTFValues(CTF* ctf, Image* number_of_extrema, Image* ctf_values) {
-    MyDebugAssertTrue(number_of_extrema->is_in_memory, "Memory not allocated");
-    MyDebugAssertTrue(ctf_values->is_in_memory, "Memory not allocated");
-    MyDebugAssertTrue(ctf_values->HasSameDimensionsAs(number_of_extrema), "Images do not have same dimensions");
-
-    int   i, j;
-    float i_logi, i_logi_sq;
-    float j_logi, j_logi_sq;
-    float current_spatial_frequency_squared;
-    float current_azimuth;
-    long  address;
-
-    address = 0;
-    for ( j = 0; j < number_of_extrema->logical_y_dimension; j++ ) {
-        j_logi    = float(j - number_of_extrema->physical_address_of_box_center_y) * number_of_extrema->fourier_voxel_size_y;
-        j_logi_sq = pow(j_logi, 2);
-        for ( i = 0; i < number_of_extrema->logical_x_dimension; i++ ) {
-            i_logi    = float(i - number_of_extrema->physical_address_of_box_center_x) * number_of_extrema->fourier_voxel_size_x;
-            i_logi_sq = pow(i_logi, 2);
-            // Where are we?
-            current_spatial_frequency_squared = j_logi_sq + i_logi_sq;
-            if ( current_spatial_frequency_squared > 0.0 ) {
-                current_azimuth = atan2(j_logi, i_logi);
-            }
-            else {
-                current_azimuth = 0.0;
-            }
-            //
-            ctf_values->real_values[address]        = ctf->Evaluate(current_spatial_frequency_squared, current_azimuth);
-            number_of_extrema->real_values[address] = ctf->ReturnNumberOfExtremaBeforeSquaredSpatialFrequency(current_spatial_frequency_squared, current_azimuth);
-            //
-            address++;
-        }
-        address += number_of_extrema->padding_jump_value;
-    }
-
-    number_of_extrema->is_in_real_space = true;
-    ctf_values->is_in_real_space        = true;
-}
